requirements.txt
django
drf_yasg
pytest
djangorestframework